#162: changed individual-detail to responsive design
