--- src/app/individual/individual-description/individual-description.component.html
@@ -1,76 +1,0 @@
-<div class="detail-label">{{ (species$ | async)?.de | translate }}</div>
-<div class="mat-body-1 clip-text" data-test-id="individual_name">{{ (individual$ | async).name }}</div>
-<div class="detail-label" *ngIf="(individual$ | async).last_observation_date">
-  <span translate>Letzte Beobachtung</span>
-  <span>, </span>
-  <span>{{ lastObservationDate$ | async }}</span>
-</div>
-<div class="phenophase-title" *ngIf="lastPhenophase$ | async">
-  <span class="phenophase-circle" [ngStyle]="{ 'background-color': lastPhenophaseColor$ | async }"></span>
-  <span>{{ (lastPhenophase$ | async)?.de | translate }}</span>
-</div>
-<div>
-  <span class="detail-label">{{ 'Von' | translate }}</span>
-  <a mat-button [routerLink]="['/profile', (individual$ | async).user]" data-test-id="individual_creator"
-    >{{ individualCreatorNickname$ | async }}
-  </a>
-  <button
-    mat-icon-button
-    class="badge"
-    matTooltip="{{
-      'PhenoRanger sind speziell ausgebildete Personen, welche Daten zu zusätzlichen Arten erfassen können. Für weitere Informationen, oder um selbst PhenoRanger zu werden, wenden Sie sich an pheno.rangers@wsl.ch.'
-        | translate
-    }}"
-    *ngIf="isRanger$ | async"
-  >
-    <mat-icon svgIcon="ranger-badge"></mat-icon>
-  </button>
-</div>
-<div class="detail-label">{{ 'Umgebung' | translate }}</div>
-<div class="detail-info-value">{{ (description$ | async)?.de | translate }}</div>
-<div class="detail-label">{{ 'Höhe' | translate }}</div>
-<div>
-  <span class="detail-info-value">{{ (individual$ | async)?.altitude | round }}</span>
-  <span class="detail-info-value-addition" translate> {{ 'm ü. M.' }}</span>
-</div>
-<div class="detail-label">{{ 'Exposition' | translate }}</div>
-<div class="detail-info-value">{{ (exposition$ | async)?.de | translate }}</div>
-<div class="detail-label">{{ 'Hangneigung' | translate }}</div>
-<div class="detail-info-value">{{ (individual$ | async)?.gradient }}%</div>
-<div class="detail-label">{{ 'Schatten' | translate }}</div>
-<div class="detail-info-value">{{ (shade$ | async)?.de | translate }}</div>
-<div class="detail-label">{{ 'Bewässerung' | translate }}</div>
-<div class="detail-info-value">{{ (irrigation$ | async)?.de | translate }}</div>
-<div class="detail-label">{{ 'Distanz zu Gebäude / Strasse' | translate }}</div>
-<div class="detail-info-value">{{ (distance$ | async)?.de | translate }}</div>
-<div class="detail-label">{{ 'Habitat' | translate }}</div>
-<div class="detail-info-value">{{ (habitat$ | async)?.de | translate }}</div>
-<div class="detail-label">{{ 'Waldtyp' | translate }}</div>
-<div class="detail-info-value">{{ (forest$ | async)?.de | translate }}</div>
-
-<div>
-  <button
-    class="navigation"
-    mat-flat-button
-    color="primary"
-    type="button"
-    id="edit-button"
-    [routerLink]="['/individuals', individualId, 'edit']"
-    *ngIf="isEditable$ | async"
-  >
-    {{ 'Objekt bearbeiten' | translate }}
-  </button>
-</div>
-<div>
-  <button
-    class="navigation"
-    mat-flat-button
-    color="accent"
-    type="button"
-    id="delete-button"
-    (click)="deleteIndividual()"
-    *ngIf="isEditable$ | async"
-  >
-    {{ 'Objekt löschen' | translate }}
-  </button>
-</div>
