--- src/app/individual/individual-description/individual-description.component.scss
@@ -1,23 +1,0 @@
-@use '~@angular/material' as mat;
-
-@import '../../../theming/typography';
-
-button.badge {
-  transform: scale(0.7);
-  width: 20px;
-}
-
-button.navigation {
-  padding-left: 40px;
-  padding-right: 40px;
-  margin-top: 0.5rem;
-  width: 12.5rem;
-}
-
-#edit-button {
-  margin-top: 4.125rem;
-}
-
-#delete-button {
-  color: #d37872;
-}
