+++ src/app/individual/individual-detail/individual-detail.component.html
@@ -1,30 +1,33 @@
-<div gdColumns="repeat(8, 1fr)" *ngIf="(detailSubject$ | async) !== null">
-  <div gdColumn="2 / span 6" gdColumn.lt-md="1 / span 8" gdColumns="repeat(6, 1fr)" class="detail-view">
+<div class="individual-detail" *ngIf="(detailSubject$ | async) !== null">
+  <div class="individual-detail__card">
     <app-individual-detail-header
-      gdColumn="1 / span 6"
-      class="detail-header"
+      class="individual-detail__header detail-header"
       [individual$]="detailSubject$"
     ></app-individual-detail-header>
-    <app-individual-subscription-button
-      *ngIf="isLoggedIn && (isOwn$ | async) === false"
-      gdColumn="1 / span 6"
-      class="individual-subscription-button"
-      [individual$]="detailSubject$"
-    ></app-individual-subscription-button>
-    <div gdColumn="1 / span 6" gdColumns="repeat(6, 1fr)" class="detail-content">
-      <app-individual-description
+    <div class="individual-detail__subscription-button" *ngIf="isLoggedIn && (isOwn$ | async) === false">
+      <app-individual-subscription-button [individual$]="detailSubject$"></app-individual-subscription-button>
+    </div>
+    <div class="individual-detail__content">
+      <!-- <div class="individual-detail__left-column"> -->
+      <app-individual-description-header
         *ngIf="detailSubject$ | async"
-        gdColumn="1 / span 2"
-        gdColumn.lt-md="1 / span 6"
-        class="detail-left"
+        class="individual-detail__description-header"
+        [individual$]="detailSubject$"
+      ></app-individual-description-header>
+      <app-individual-description-basic-info
+        *ngIf="detailSubject$ | async"
+        class="individual-detail__basic-info"
+        [individual$]="detailSubject$"
+      ></app-individual-description-basic-info>
+      <app-individual-description-buttons
+        *ngIf="detailSubject$ | async"
+        class="individual-detail__buttons"
         [individual$]="detailSubject$"
         [individualId]="detailId"
         [isEditable$]="isEditable$"
-      ></app-individual-description>
+      ></app-individual-description-buttons>
       <app-individual-observation-view
-        gdColumn="3 / span 4"
-        gdColumn.lt-md="1 / span 6"
-        class="detail-right"
+        class="individual-detail__observation-view"
         [individual$]="detailSubject$"
         [individualId]="detailId"
         [isEditable$]="isEditable$"
