+++ src/app/individual/individual-edit/individual-edit.component.scss
@@ -1,3 +1,26 @@
-.detail-view {
-  background-color: white;
+@import '../../../theming/mixins';
+
+.individual-edit {
+  @include detail-view;
+
+  &__card {
+    @include detail-card;
+  }
+
+  &__header {
+  }
+
+  &__content {
+    @include detail-view-content;
+  }
+
+  &__left-content {
+    @include default-left-right-padding;
+    @include left-column;
+  }
+
+  &__right-content {
+    @include default-left-right-padding;
+    @include complete-right-column;
+  }
 }
