+++ src/theming/_constants.scss
@@ -1,0 +1,7 @@
+// Breakpoints
+$min-desktop-width: 959px;
+$min-laptop-width: 750px;
+
+body {
+  --default-padding: 20px;
+}
