+++ src/theming/_overwrite-mat.scss
@@ -11,6 +11,7 @@ button[mat-flat-button] {
 }
 
 /* fix for https://github.com/angular/components/issues/23784 */
+/* stylelint-disable selector-type-no-unknown */
 #phaenonet-body mat-card-actions.phaenonet-overwrite-mat-card-actions-center {
   display: flex;
   justify-content: center;
@@ -18,4 +19,13 @@ button[mat-flat-button] {
 
 #phaenonet-body mat-card-subtitle {
   @include mat.typography-level($phaenonet-typography, 'body-2');
+}
+/* stylelint-enable selector-type-no-unknown */
+
+#phaenonet-body .mat-form-field-label {
+  color: #d8d8d8;
+}
+
+#phaenonet-body .mat-form-field-underline {
+  background-color: #f7f5f2;
 }
